Added a command to create a backup
